+++ backend/src/main/java/org/ftclub/cabinet/oauth/domain/OauthResult.java
@@ -1,0 +1,17 @@
+package org.ftclub.cabinet.oauth.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class OauthResult {
+
+	private Long userId;
+	private String roles;
+	private String redirectionUrl;
+
+	public boolean hasRole(String role) {
+		return role != null && role.contains(roles);
+	}
+}
