+++ backend/src/main/java/org/ftclub/cabinet/oauth/service/OauthFacadeService.java
@@ -6,10 +6,10 @@ import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.ftclub.cabinet.admin.auth.service.AdminAuthService;
-import org.ftclub.cabinet.auth.domain.CustomOauth2User;
-import org.ftclub.cabinet.auth.domain.FtOauthProfile;
-import org.ftclub.cabinet.auth.domain.OauthResult;
 import org.ftclub.cabinet.auth.service.AuthPolicyService;
+import org.ftclub.cabinet.oauth.domain.CustomOauth2User;
+import org.ftclub.cabinet.oauth.domain.FtOauthProfile;
+import org.ftclub.cabinet.oauth.domain.OauthResult;
 import org.ftclub.cabinet.security.SecurityPathPolicyService;
 import org.ftclub.cabinet.user.domain.User;
 import org.ftclub.cabinet.user.service.UserFacadeService;
