+++ backend/src/main/java/org/ftclub/cabinet/oauth/service/OauthLinkFacadeService.java
@@ -3,18 +3,18 @@ package org.ftclub.cabinet.oauth.service;
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.ftclub.cabinet.auth.domain.CustomOauth2User;
-import org.ftclub.cabinet.auth.domain.FtOauthProfile;
-import org.ftclub.cabinet.auth.domain.OauthResult;
 import org.ftclub.cabinet.auth.service.ApplicationTokenManager;
 import org.ftclub.cabinet.auth.service.AuthPolicyService;
 import org.ftclub.cabinet.auth.service.CookieService;
-import org.ftclub.cabinet.security.exception.SpringSecurityException;
 import org.ftclub.cabinet.dto.UserInfoDto;
 import org.ftclub.cabinet.exception.ExceptionStatus;
 import org.ftclub.cabinet.jwt.domain.JwtTokenConstants;
 import org.ftclub.cabinet.jwt.service.JwtService;
+import org.ftclub.cabinet.oauth.domain.CustomOauth2User;
+import org.ftclub.cabinet.oauth.domain.FtOauthProfile;
 import org.ftclub.cabinet.oauth.domain.OauthLink;
+import org.ftclub.cabinet.oauth.domain.OauthResult;
+import org.ftclub.cabinet.security.exception.SpringSecurityException;
 import org.ftclub.cabinet.user.domain.User;
 import org.ftclub.cabinet.user.service.UserFacadeService;
 import org.ftclub.cabinet.user.service.UserQueryService;
