+++ backend/src/main/java/org/ftclub/cabinet/openpublic/controller/LoginController.java
@@ -5,9 +5,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.ftclub.cabinet.alarm.config.AlarmProperties;
-import org.ftclub.cabinet.auth.domain.OauthResult;
-import org.ftclub.cabinet.auth.service.AuthenticationService;
+import org.ftclub.cabinet.auth.service.AuthFacadeService;
 import org.ftclub.cabinet.log.Logging;
+import org.ftclub.cabinet.oauth.domain.OauthResult;
 import org.ftclub.cabinet.oauth.service.OauthFacadeService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class LoginController {
 
 	private final AlarmProperties alarmProperties;
-	private final AuthenticationService authenticationService;
+	private final AuthFacadeService authFacadeService;
 	private final OauthFacadeService oauthFacadeService;
 
 	@GetMapping("")
@@ -29,7 +29,7 @@ public class LoginController {
 		if (!alarmProperties.getIsProduction()) {
 			OauthResult result = oauthFacadeService.handlePublicLogin(username);
 
-			authenticationService.processAuthentication(request, response, result, "Temporary");
+			authFacadeService.processAuthentication(request, response, result, "Temporary");
 		}
 	}
 }
