+++ backend/src/main/java/org/ftclub/cabinet/utils/leave/absence/LeaveAbsenceManager.java
@@ -1,34 +1,34 @@
 package org.ftclub.cabinet.utils.leave.absence;
 
+import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.ftclub.cabinet.auth.domain.FtProfile;
 import org.ftclub.cabinet.auth.domain.FtRole;
 import org.ftclub.cabinet.auth.service.ApplicationTokenManager;
-import org.ftclub.cabinet.auth.service.UserOauthService;
 import org.ftclub.cabinet.lent.service.LentFacadeService;
+import org.ftclub.cabinet.oauth.domain.FtOauthProfile;
+import org.ftclub.cabinet.oauth.service.OauthProfileService;
 import org.ftclub.cabinet.user.service.UserCommandService;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.HttpClientErrorException;
-
-import java.time.LocalDateTime;
 
 @Component
 @RequiredArgsConstructor
 @Log4j2
 public class LeaveAbsenceManager {
 
-	private final UserOauthService userOauthService;
 	private final ApplicationTokenManager tokenManager;
 	private final LentFacadeService lentFacadeService;
 	private final UserCommandService userCommandService;
+	private final OauthProfileService oauthProfileService;
 
 	public void handleLeaveAbsence(Long userId, String name) {
 		log.info("called handleLeaveAbsence {} {}", userId, name);
 		try {
-			FtProfile ftProfile = userOauthService.getProfileByIntraName(tokenManager.getFtAccessToken(), name);
-			if (ftProfile.getRole().equals(FtRole.INACTIVE)) {
+			FtOauthProfile ftProfile = oauthProfileService.getProfileByIntraName(
+					tokenManager.getFtAccessToken(), name);
+			if (ftProfile.hasRole((FtRole.INACTIVE))) {
 				lentFacadeService.endUserLent(userId, null);
 			}
 		} catch (HttpClientErrorException e) {
